refactor(web): W10 slice 4 — orphan directories merge into their real homes (BI-AB17E1A8)
The pass §3.1-b table: "lib/workspace/ vs lib/workspace-home/, lib/edge/
vs lib/edge-node/, lib/platform/ vs lib/platform-runtime/ vs
lib/platform-config/ — one-file orphans beside their real home".

- lib/workspace/command-center(+test) -> lib/workspace-home/ (the
  16-entry workspace-home domain already imports it via platform-loader);
- lib/edge/{metrics-cache,metrics-rate-limit} -> lib/edge-node/ (both are
  edge-NODE telemetry per their spec header, 2026-05-19 edge-node
  network-telemetry design §6.2);
- lib/platform-config/platform-dev-config-core(+test) -> lib/platform/
  (single-module dir; platform-dev config helpers beside the platform
  version/image-version modules).

DEFERRED (documented, not moved): the lib/platform vs lib/platform-runtime
merge. lib/platform-runtime is wired into the scripts-side substrate
chain (scripts/lib/derived-artifacts-registry.mjs, capability compose
profile guards, measure-platform-substrate, capability-state-hash) —
folding it is a cascading change beyond mechanical import rewrites, out
of an XL move-only wave per the seam brief.

Move-only via git mv; 10 import specifiers rewritten; the three dissolved
directories are gone. module-size baseline entry follows
workspace/command-center.ts (--update; diff is exactly that one path).

Characterization: workspace-home + edge-node + platform + workspace/
components + edge metrics route vitest 56 files / 563 tests green;
per-file test-count diff vs pre-move baseline EMPTY modulo renamed
prefixes. Application-boundary guard unchanged: 8 contexts, 531
cross-context imports, 63 exceptions.

Docs-Impact-Decision: internal namespace hygiene, no user-facing change.
Design-Grounding-Decision: grounded in
docs/architecture/2026-08-16-simplify-and-strengthen-architecture-pass.md
§3.1-b + §5 W10 (BI-AB17E1A8); substrate reviewed — workspace-home
platform-loader (existing importer), edge-node telemetry spec lineage,
platform-runtime scripts-side coupling (deferral evidence).
Process-Spine-Decision: executes a documented architecture-pass move
(§3.1-b one-file orphan merges); mechanical git mv + import rewrites.
