+++ scripts/module-size-baseline.txt
@@ -69,5 +69,5 @@ apps/web/lib/work-capsules/work-capsule-store.test.ts	1085
 apps/web/lib/work-capsules/work-capsule-store.ts	1034
 apps/web/lib/workbooks/platform-tables.ts	938
 apps/web/lib/workforce/calendar-data.ts	1134
-apps/web/lib/workspace/command-center.ts	1118
+apps/web/lib/workspace-home/command-center.ts	1118
 packages/db/src/wiki-store.test.ts	824
